chore(reorg): rename basic input
