+++ src/data-workspace/inputs/index.js
@@ -1,6 +1,6 @@
+export * from './basic-input.js'
+export * from './boolean-radios.js'
 export * from './file-inputs.js'
 export * from './long-text.js'
-export * from './inputs.js'
-export * from './true-only-checkbox.js'
-export * from './boolean-radios.js'
 export * from './option-set.js'
+export * from './true-only-checkbox.js'
